feat: Move NotifyConfigDTO to Explorer module
