+++ core/android/src/main/kotlin/com/walletconnect/android/internal/common/explorer/data/network/model/TypeDTO.kt
@@ -1,6 +1,6 @@
 @file:JvmSynthetic
 
-package com.walletconnect.notify.data.wellknown.config
+package com.walletconnect.android.internal.common.explorer.data.network.model
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
@@ -9,6 +9,8 @@ import com.squareup.moshi.JsonClass
 internal data class TypeDTO(
     @Json(name = "name")
     val name: String,
+    @Json(name = "id")
+    val id: String,
     @Json(name = "description")
-    val description: String
+    val description: String,
 )
